generator/requirements.txt
eth_abi==5.1.0
Flask==3.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
web3==7.5.0
